requirements.txt
asgiref==3.8.1
confluent-kafka==2.3.0
Django==4.2.13
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
kafka-python @ git+https://github.com/dpkp/kafka-python.git@a6d0579d3cadd3826dd364b01bc12a2173139abc
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.2.0
sqlparse==0.4.4
StrEnum==0.4.15
uritemplate==4.1.1
